chore: upgrade to sveltekit libary
