+++ .github/workflows/release.yml
@@ -13,7 +13,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v3
         with:
-         fetch-depth: 0
+          fetch-depth: 0
       - name: Setup Node.js 16
         uses: actions/setup-node@v3
         with:
