+++ .gitignore
@@ -3,3 +3,11 @@ dist
 node_modules
 public/bundle.*
 coverage
+/build
+/.svelte-kit
+/package
+.env
+.env.*
+!.env.example
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
