+++ .npmrc
@@ -1,1 +1,2 @@
 scripts-prepend-node-path=true
+engine-strict=true
