--- jest.config.js
@@ -1,16 +1,0 @@
-module.exports = {
-  transform: {
-    '^.+\\.m?js$': 'babel-jest',
-    "^.+\\.svelte$": ["svelte-jester", { "preprocess": true }]
-  },
-  transformIgnorePatterns: [
-    '/node_modules/(?!(lodash-es|svelte-writable-derived)).+\\.m?js$',
-  ],
-  moduleFileExtensions: ['js', 'svelte', 'mjs'],
-  roots: ['<rootDir>/tests'],
-  bail: false,
-  verbose: false,
-  collectCoverage: true,
-  collectCoverageFrom: ['src/components/**/*.{js,svelte}', 'src/utils.js'],
-  setupFilesAfterEnv: ['@testing-library/jest-dom/extend-expect'],
-};
