+++ package.json
@@ -1,109 +1,94 @@
 {
-  "name": "@thatconference/sveltejs-forms",
-  "description": "Declarative forms for Svelte",
-  "homepage": "https://mdauner.github.io/sveltejs-forms/",
-  "version": "0.0.0-development",
-  "license": "MIT",
-  "repository": "github:thatconference/sveltejs-forms",
-  "svelte": "src/components/components.module.js",
-  "module": "dist/index.min.mjs",
-  "main": "dist/index.min.js",
-  "devDependencies": {
-    "@babel/core": "~7.11.6",
-    "@babel/plugin-syntax-dynamic-import": "~7.8.3",
-    "@babel/plugin-transform-runtime": "~7.11.5",
-    "@babel/preset-env": "~7.11.5",
-    "@commitlint/cli": "~8.3.5",
-    "@commitlint/config-conventional": "~8.3.4",
-    "@testing-library/dom": "^7.24.1",
-    "@testing-library/jest-dom": "~5.11.4",
-    "@testing-library/svelte": "~3.0.0",
-    "@testing-library/user-event": "^12.1.3",
-    "@types/jest": "~25.1.3",
-    "autoprefixer": "~9.7.4",
-    "babel-jest": "~26.3.0",
-    "cz-conventional-changelog": "~3.1.0",
-    "eslint": "~6.8.0",
-    "eslint-config-prettier": "~6.10.00",
-    "eslint-plugin-jest": "~23.8.1",
-    "eslint-plugin-svelte3": "~2.7.3",
-    "husky": "~4.2.3",
-    "jest": "~25.1.0",
-    "lint-staged": "~10.0.8",
-    "node-sass": "~7.0.3",
-    "npm-run-all": "~4.1.5",
-    "postcss": "~7.0.27",
-    "postcss-load-config": "~2.1.0",
-    "prettier": "~1.19.1",
-    "prettier-plugin-svelte": "~0.7.0",
-    "rollup": "~1.32.0",
-    "rollup-plugin-babel": "~4.3.3",
-    "rollup-plugin-commonjs": "~10.1.0",
-    "rollup-plugin-livereload": "~1.0.4",
-    "rollup-plugin-node-resolve": "~5.2.0",
-    "rollup-plugin-svelte": "~5.1.1",
-    "rollup-plugin-terser": "~5.2.0",
-    "semantic-release": "^19.0.5",
-    "sirv-cli": "~0.4.5",
-    "svelte": "~3.20.0",
-    "svelte-jester": "^1.1.5",
-    "svelte-preprocess": "~3.4.0",
-    "yup": "~0.28.2"
-  },
-  "peerDependencies": {
-    "svelte": "~3.20.0",
-    "yup": "~0.28.2"
-  },
-  "dependencies": {
-    "lodash-es": "~4.17.15",
-    "svelte-writable-derived": "~2.0.1"
-  },
-  "scripts": {
-    "prettier": "prettier --check '**/*.{svelte, html, css, scss, stylus, js, ts, json, yml, md}' --plugin-search-dir=.",
-    "lint": "eslint --color './**/*.{js,svelte}'",
-    "test": "jest --verbose",
-    "autobuild": "rollup -c -w",
-    "start:dev": "sirv public --single --dev",
-    "dev": "run-p start:dev autobuild",
-    "build": "rollup -c",
-    "prepublishOnly": "npm run build",
-    "semantic-release": "semantic-release"
-  },
-  "browserslist": [
-    "defaults"
-  ],
-  "keywords": [
-    "svelte",
-    "forms",
-    "validation",
-    "javascript"
-  ],
-  "files": [
-    "src",
-    "dist"
-  ],
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged",
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-    }
-  },
-  "lint-staged": {
-    "*.{svelte,js}": [
-      "eslint --fix",
-      "git add"
-    ],
-    "*.{svelte, html, css, scss, stylus, js, ts, json, yml, md}": [
-      "prettier --write --plugin-search-dir=.",
-      "git add"
-    ]
-  },
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
-  },
-  "publishConfig": {
-    "access": "public"
-  }
+	"name": "@thatconference/sveltejs-forms",
+	"description": "Declarative forms for Svelte",
+	"homepage": "https://mdauner.github.io/sveltejs-forms/",
+	"version": "0.0.0-development",
+	"license": "MIT",
+	"repository": "github:thatconference/sveltejs-forms",
+	"svelte": "index.js",
+	"type": "module",
+	"devDependencies": {
+		"@commitlint/cli": "^17.1.2",
+		"@commitlint/config-conventional": "^17.1.0",
+		"@sveltejs/adapter-auto": "^1.0.0-next.90",
+		"@sveltejs/kit": "^1.0.0-next.587",
+		"@sveltejs/package": "^1.0.0-next.6",
+		"@testing-library/jest-dom": "~5.11.4",
+		"@testing-library/svelte": "^3.2.2",
+		"@testing-library/user-event": "^14.4.3",
+		"@types/jest": "~25.1.3",
+		"cz-conventional-changelog": "^3.3.0",
+		"eslint": "^8.28.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-jest": "^27.2.1",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"husky": "^4.2.3",
+		"jsdom": "^21.1.1",
+		"lint-staged": "^13.2.0",
+		"prettier": "^2.8.0",
+		"prettier-plugin-svelte": "^2.8.1",
+		"sass": "^1.60.0",
+		"semantic-release": "^19.0.5",
+		"svelte": "^3.52.0",
+		"svelte-jester": "^2.3.2",
+		"tslib": "^2.4.1",
+		"typescript": "^4.9.3",
+		"vite": "^4.0.0",
+		"vitest": "^0.25.3",
+		"yup": "~0.28.2"
+	},
+	"peerDependencies": {
+		"svelte": "^3.52.0",
+		"yup": "~0.28.2"
+	},
+	"dependencies": {
+		"lodash-es": "^4.17.21",
+		"svelte-writable-derived": "^3.0.1"
+	},
+	"scripts": {
+		"dev": "vite dev",
+		"build": "svelte-kit sync && svelte-package",
+		"lint": "prettier --plugin-search-dir . --check . && eslint .",
+		"format": "prettier --plugin-search-dir . --write .",
+		"test": "vitest run",
+		"prepublishOnly": "echo 'Did you mean to publish `./package/`, instead of `./`?' && exit 1",
+		"semantic-release": "semantic-release"
+	},
+	"browserslist": [
+		"defaults"
+	],
+	"keywords": [
+		"svelte",
+		"forms",
+		"validation",
+		"javascript"
+	],
+	"files": [
+		"src",
+		"dist"
+	],
+	"husky": {
+		"hooks": {
+			"pre-commit": "lint-staged",
+			"commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+		}
+	},
+	"lint-staged": {
+		"*.{svelte,js}": [
+			"eslint --fix",
+			"git add"
+		],
+		"*.{svelte, html, css, scss, stylus, js, ts, json, yml, md}": [
+			"prettier --write --plugin-search-dir=.",
+			"git add"
+		]
+	},
+	"config": {
+		"commitizen": {
+			"path": "./node_modules/cz-conventional-changelog"
+		}
+	},
+	"publishConfig": {
+		"access": "public"
+	}
 }
