--- public/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf8" />
-    <meta name="viewport" content="width=device-width" />
-
-    <title>Svelte component</title>
-
-    <link rel="stylesheet" href="bundle.css" />
-  </head>
-
-  <body>
-    <script src="bundle.js"></script>
-  </body>
-</html>
