--- src/main.js
@@ -1,7 +1,0 @@
-import App from './App.svelte';
-
-const app = new App({
-  target: document.body,
-});
-
-export default app;
