--- tests/Input/TestApp.svelte
@@ -1,26 +1,0 @@
-<script>
-  import { Form, Input } from '../../src/components/components.module.js';
-
-  export let onSubmit = () => {};
-  export let schema = null;
-  export let initialValues = {};
-  export let validateOnBlur = true;
-  export let validateOnChange = true;
-  export let form = null;
-  export let multiline = false;
-  export let label = '';
-  export let disabled = false;
-</script>
-
-<Form
-  {schema}
-  {initialValues}
-  {validateOnBlur}
-  {validateOnChange}
-  on:submit={onSubmit}
-  let:isSubmitting
-  bind:this={form}>
-  <Input name="email" {label} placeholder="Email" {multiline} {disabled} />
-
-  <button type="submit" disabled={isSubmitting}>Sign in</button>
-</Form>
